Reduce main thread blocking for `get_feature`
If a get_feature expression function is called in a background thread
it no longer needs to block the main thread within each `eval` to get
access to the vector layer in most cases. This has caused performance
problems and even freeze/crash (through connection pool depletion).

If the `layer` parameter is static (e.g. by directly providing a layer name)
it will now be resolved to a QgsVectorLayerFeatureSoruce in the expression
preparation phase which is cheap and happens on the main thread already.
Later on the thread can work with this source indepentently of the main
thread.